pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openagent"
version = "0.1.0"
description = "Zero-cost, offline-first, open-source AI agent"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2",
    "PyMuPDF>=1.23",
    "python-docx>=1.0",
    "Pillow>=10.0",
    "pytesseract>=0.3",
    "duckduckgo-search>=6.0",
    "torch>=2.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4", "pytest-asyncio>=0.21"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
